Add denomination service
