+++ cash-register/src/app/app.module.ts
@@ -3,7 +3,6 @@ import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
 import { CalculatorComponent } from './calculator/calculator.component';
-import { Denominations } from './denominations';
 
 @NgModule({
   declarations: [
