+++ cash-register/src/app/denominations.service.ts
@@ -1,0 +1,24 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class DenominationsService {
+
+  constructor() { }
+
+  getDenominations() {
+    var denominations = [
+      {name: "twenty", plural: "twenties", value: 20.00},
+      {name: "ten", plural: "tens", value: 10.00},
+      {name: "five", plural: "fives", value: 5.00},
+      {name: "one", plural: "ones", value: 1.00},
+      {name: "quarter", plural: "quarters", value: 0.25},
+      {name: "dime", plural: "dimes", value: 0.10},
+      {name: "nickle", plural: "nickles", value: 0.05},
+      {name: "penny", plural: "pennies", value: 0.01}
+    ];
+    return denominations
+  }
+
+}
